Reorganize logger
