+++ llumnix/backends/vllm/executor.py
@@ -30,7 +30,7 @@ from vllm.sequence import Logprob, SequenceOutput, SequenceGroupOutput, SamplerO
 from vllm.config import _GB
 
 from llumnix.internal_config import MigrationConfig
-from llumnix.logger import init_logger
+from llumnix.logging.logger import init_logger
 from llumnix.backends.vllm.utils import get_cache_block_size
 from llumnix.backends.profiling import LatencyMemData, SimCacheConfig, model_prefill, model_decode, _pad_to_alignment
