+++ llumnix/backends/vllm/migration_backend.py
@@ -20,7 +20,7 @@ import ray.util.collective as col
 from vllm.worker.cache_engine import CacheEngine
 from llumnix.internal_config import MigrationConfig
 from llumnix.backends.migration_backend_interface import MigrationBackendBase
-from llumnix.logger import init_logger
+from llumnix.logging.logger import init_logger
 
 logger = init_logger(__name__)
