+++ llumnix/entrypoints/bladellm/utils.py
@@ -14,7 +14,7 @@
 from blade_llm.service.args import ServingArgs
 
 from llumnix.arg_utils import EntrypointsArgs, ManagerArgs
-from llumnix.logger import init_logger
+from llumnix.logging.logger import init_logger
 
 logger = init_logger(__name__)
