+++ llumnix/global_scheduler/dispatch_scheduler.py
@@ -15,7 +15,7 @@ from typing import Dict, List, Set
 from abc import ABC, abstractmethod
 import random
 
-from llumnix.logger import init_logger
+from llumnix.logging.logger import init_logger
 from llumnix.instance_info import InstanceLoadCalculator, InstanceInfo
 
 logger = init_logger(__name__)
