+++ llumnix/global_scheduler/scaling_scheduler.py
@@ -16,7 +16,7 @@ from abc import ABC, abstractmethod
 from enum import Enum
 import numpy as np
 
-from llumnix.logger import init_logger
+from llumnix.logging.logger import init_logger
 from llumnix.instance_info import InstanceInfo, InstanceLoadCalculator
 
 logger = init_logger(__name__)
