+++ llumnix/logging/logger.py
@@ -43,7 +43,7 @@ _DATE_FORMAT = "%m-%d %H:%M:%S"
 DEFAULT_LOGGING_CONFIG = {
     "formatters": {
         "llumnix": {
-            "class": "llumnix.logging_utils.NewLineFormatter",
+            "class": "llumnix.logging.NewLineFormatter",
             "datefmt": _DATE_FORMAT,
             "format": _FORMAT,
         },
@@ -163,67 +163,3 @@ def init_logger(name: str) -> _LlumnixLogger:
 _configure_llumnix_root_logger()
 
 logger = init_logger(__name__)
-
-
-# pylint: disable=unused-argument,inconsistent-return-statements
-def _trace_calls(log_path, root_dir, frame, event, arg=None):
-    if event in ['call', 'return']:
-        # Extract the filename, line number, function name, and the code object
-        filename = frame.f_code.co_filename
-        lineno = frame.f_lineno
-        func_name = frame.f_code.co_name
-        if not filename.startswith(root_dir):
-            # only log the functions in the llumnix root_dir
-            return
-        # Log every function call or return
-        try:
-            last_frame = frame.f_back
-            if last_frame is not None:
-                last_filename = last_frame.f_code.co_filename
-                last_lineno = last_frame.f_lineno
-                last_func_name = last_frame.f_code.co_name
-            else:
-                # initial frame
-                last_filename = ""
-                last_lineno = 0
-                last_func_name = ""
-            # pylint: disable=unspecified-encoding
-            with open(log_path, 'a') as f:
-                ts = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f")
-                if event == 'call':
-                    f.write(f"{ts} Call to"
-                            f" {func_name} in {filename}:{lineno}"
-                            f" from {last_func_name} in {last_filename}:"
-                            f"{last_lineno}\n")
-                else:
-                    f.write(f"{ts} Return from"
-                            f" {func_name} in {filename}:{lineno}"
-                            f" to {last_func_name} in {last_filename}:"
-                            f"{last_lineno}\n")
-        except NameError:
-            # modules are deleted during shutdown
-            pass
-    return partial(_trace_calls, log_path, root_dir)
-
-
-def enable_trace_function_call(log_file_path: str,
-                               root_dir: Optional[str] = None):
-    """
-    Enable tracing of every function call in code under `root_dir`.
-    This is useful for debugging hangs or crashes.
-    `log_file_path` is the path to the log file.
-    `root_dir` is the root directory of the code to trace. If None, it is the
-    llumnix root directory.
-
-    Note that this call is thread-level, any threads calling this function
-    will have the trace enabled. Other threads will not be affected.
-    """
-    logger.warning(
-        "LLUMNIX_TRACE_FUNCTION is enabled. It will record every"
-        " function executed by Python. This will slow down the code. It "
-        "is suggested to be used for debugging hang or crashes only.")
-    logger.info("Trace frame log is saved to %s", log_file_path)
-    if root_dir is None:
-        # by default, this is the llumnix root directory
-        root_dir = os.path.dirname(os.path.dirname(__file__))
-    sys.settrace(partial(_trace_calls, log_file_path, root_dir))
