+++ llumnix/manager.py
@@ -25,7 +25,7 @@ from ray.util.state import list_placement_groups, list_actors
 from ray.util.placement_group import PlacementGroup
 
 from llumnix.llumlet.llumlet import Llumlet
-from llumnix.logger import init_logger
+from llumnix.logging.logger import init_logger
 from llumnix.global_scheduler.global_scheduler import GlobalScheduler
 from llumnix.global_scheduler.migration_scheduler import PairMigrationConstraints
 from llumnix.global_scheduler.migration_filter import CustomFilter
@@ -149,7 +149,7 @@ class Manager:
 
     def __repr__(self):
         # Customizing prefixes for Actor logs.
-        return f"{self.__class__.__name__}(node_id={self.node_id[:5]})"
+        return f"{self.__class__.__name__}(nid={self.node_id[:5]})"
 
     async def generate(self, request_id: str, server_info: ServerInfo, *args, **kwargs,) -> None:
         while self.num_instances == 0:
