+++ tests/unit_test/logging/test_logger.py
@@ -27,33 +27,9 @@ from uuid import uuid4
 
 import pytest
 
-from llumnix.logger import (_DATE_FORMAT, _FORMAT, _configure_llumnix_root_logger,
-                            enable_trace_function_call, init_logger)
-from llumnix.logging_utils import NewLineFormatter
-
-
-def f1(x):
-    return f2(x)
-
-
-def f2(x):
-    return x
-
-
-def test_trace_function_call():
-    # pylint: disable=unused-variable
-    fd, path = tempfile.mkstemp()
-    cur_dir = os.path.dirname(__file__)
-    enable_trace_function_call(path, cur_dir)
-    f1(1)
-    # pylint: disable=unspecified-encoding
-    with open(path) as f:
-        content = f.read()
-
-    assert "f1" in content
-    assert "f2" in content
-    sys.settrace(None)
-    os.remove(path)
+from llumnix.logging.logger import (_DATE_FORMAT, _FORMAT, _configure_llumnix_root_logger,
+                                    init_logger)
+from llumnix.logging import NewLineFormatter
 
 
 def test_default_llumnix_root_logger_configuration():
@@ -77,8 +53,8 @@ def test_default_llumnix_root_logger_configuration():
     assert formatter.datefmt == _DATE_FORMAT
 
 
-@patch("llumnix.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
-@patch("llumnix.logger.LLUMNIX_LOGGING_CONFIG_PATH", None)
+@patch("llumnix.logging.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
+@patch("llumnix.logging.logger.LLUMNIX_LOGGING_CONFIG_PATH", None)
 def test_descendent_loggers_depend_on_and_propagate_logs_to_root_logger():
     """This test presumes that LLUMNIX_CONFIGURE_LOGGING (default: True) and
     LLUMNIX_LOGGING_CONFIG_PATH (default: None) are not configured and default
@@ -106,21 +82,21 @@ def test_descendent_loggers_depend_on_and_propagate_logs_to_root_logger():
     assert log_record.levelno == logging.INFO
 
 
-@patch("llumnix.logger.LLUMNIX_CONFIGURE_LOGGING", 0)
-@patch("llumnix.logger.LLUMNIX_LOGGING_CONFIG_PATH", None)
+@patch("llumnix.logging.logger.LLUMNIX_CONFIGURE_LOGGING", 0)
+@patch("llumnix.logging.logger.LLUMNIX_LOGGING_CONFIG_PATH", None)
 def test_logger_configuring_can_be_disabled():
     """This test calls _configure_llumnix_root_logger again to test custom logging
     config behavior, however mocks are used to ensure no changes in behavior or
     configuration occur."""
 
-    with patch("llumnix.logger.dictConfig") as dict_config_mock:
+    with patch("llumnix.logging.logger.dictConfig") as dict_config_mock:
         _configure_llumnix_root_logger()
     dict_config_mock.assert_not_called()
 
 
-@patch("llumnix.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
+@patch("llumnix.logging.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
 @patch(
-    "llumnix.logger.LLUMNIX_LOGGING_CONFIG_PATH",
+    "llumnix.logging.logger.LLUMNIX_LOGGING_CONFIG_PATH",
     "/if/there/is/a/file/here/then/you/did/this/to/yourself.json",
 )
 def test_an_error_is_raised_when_custom_logging_config_file_does_not_exist():
@@ -133,7 +109,7 @@ def test_an_error_is_raised_when_custom_logging_config_file_does_not_exist():
     assert "File does not exist" in str(ex_info)
 
 
-@patch("llumnix.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
+@patch("llumnix.logging.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
 def test_an_error_is_raised_when_custom_logging_config_is_invalid_json():
     """This test calls _configure_llumnix_root_logger again to test custom logging
     config behavior, however it fails before any change in behavior or
@@ -141,7 +117,7 @@ def test_an_error_is_raised_when_custom_logging_config_is_invalid_json():
     with NamedTemporaryFile(encoding="utf-8", mode="w") as logging_config_file:
         logging_config_file.write("---\nloggers: []\nversion: 1")
         logging_config_file.flush()
-        with patch("llumnix.logger.LLUMNIX_LOGGING_CONFIG_PATH",
+        with patch("llumnix.logging.logger.LLUMNIX_LOGGING_CONFIG_PATH",
                    logging_config_file.name):
             with pytest.raises(JSONDecodeError) as ex_info:
                 _configure_llumnix_root_logger()
@@ -149,7 +125,7 @@ def test_an_error_is_raised_when_custom_logging_config_is_invalid_json():
             assert "Expecting value" in str(ex_info)
 
 
-@patch("llumnix.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
+@patch("llumnix.logging.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
 @pytest.mark.parametrize("unexpected_config", (
     "Invalid string",
     [{
@@ -166,7 +142,7 @@ def test_an_error_is_raised_when_custom_logging_config_is_unexpected_json(
     with NamedTemporaryFile(encoding="utf-8", mode="w") as logging_config_file:
         logging_config_file.write(json.dumps(unexpected_config))
         logging_config_file.flush()
-        with patch("llumnix.logger.LLUMNIX_LOGGING_CONFIG_PATH",
+        with patch("llumnix.logging.logger.LLUMNIX_LOGGING_CONFIG_PATH",
                    logging_config_file.name):
             with pytest.raises(ValueError) as ex_info:
                 _configure_llumnix_root_logger()
@@ -174,7 +150,7 @@ def test_an_error_is_raised_when_custom_logging_config_is_unexpected_json(
             assert "Invalid logging config. Expected Dict, got" in str(ex_info)
 
 
-@patch("llumnix.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
+@patch("llumnix.logging.logger.LLUMNIX_CONFIGURE_LOGGING", 1)
 def test_custom_logging_config_is_parsed_and_used_when_provided():
     """This test calls _configure_llumnix_root_logger again to test custom logging
     config behavior, however mocks are used to ensure no changes in behavior or
@@ -191,14 +167,14 @@ def test_custom_logging_config_is_parsed_and_used_when_provided():
     with NamedTemporaryFile(encoding="utf-8", mode="w") as logging_config_file:
         logging_config_file.write(json.dumps(valid_logging_config))
         logging_config_file.flush()
-        with patch("llumnix.logger.LLUMNIX_LOGGING_CONFIG_PATH",
+        with patch("llumnix.logging.logger.LLUMNIX_LOGGING_CONFIG_PATH",
                    logging_config_file.name), patch(
-                       "llumnix.logger.dictConfig") as dict_config_mock:
+                       "llumnix.logging.logger.dictConfig") as dict_config_mock:
             _configure_llumnix_root_logger()
             dict_config_mock.assert_called_with(valid_logging_config)
 
 
-@patch("llumnix.logger.LLUMNIX_CONFIGURE_LOGGING", 0)
+@patch("llumnix.logging.logger.LLUMNIX_CONFIGURE_LOGGING", 0)
 def test_custom_logging_config_causes_an_error_if_configure_logging_is_off():
     """This test calls _configure_llumnix_root_logger again to test custom logging
     config behavior, however mocks are used to ensure no changes in behavior or
@@ -214,7 +190,7 @@ def test_custom_logging_config_causes_an_error_if_configure_logging_is_off():
     with NamedTemporaryFile(encoding="utf-8", mode="w") as logging_config_file:
         logging_config_file.write(json.dumps(valid_logging_config))
         logging_config_file.flush()
-        with patch("llumnix.logger.LLUMNIX_LOGGING_CONFIG_PATH",
+        with patch("llumnix.logging.logger.LLUMNIX_LOGGING_CONFIG_PATH",
                    logging_config_file.name):
             with pytest.raises(RuntimeError) as ex_info:
                 _configure_llumnix_root_logger()
